+++ src/main/java/sayTheSpire/buffers/LeaderboardBuffer.java
@@ -17,7 +17,7 @@ public class LeaderboardBuffer extends Buffer {
         this.context.put("rank", entry.rank);
         this.context.put("name", entry.name);
         if (entry.isTime) {
-            this.context.put("score", CharStat.formatHMSM(entry.score));
+            this.context.put("date", CharStat.formatHMSM(entry.score));
             this.addLocalized("content.timeEntry");
         } else {
             this.context.put("score", entry.score);